arkrpc: move tree path conversion into treeconv
Move the tree path conversion helpers out of the base `arkrpc` package and
into a dedicated `arkrpc/treeconv` package. The generated RPC package should
stay as lightweight as possible, because pulling `lib/tree` into it fans a
large dependency tree out into every consumer.

This keeps the incoming-metadata callers working while giving the swap daemon
platform a much narrower import surface to lint and build against. I also
updated the local package docs so future callers know to reach for
`arkrpc/treeconv` when they need to cross the proto/tree boundary.
