+++ arkrpc/AGENTS.md
@@ -3,16 +3,14 @@
 ## Purpose
 
 Server-side gRPC service definitions (ArkService, IndexerService) with generated
-Go stubs, plus hand-written conversion utilities for domain types.
+Go stubs. Tree-path conversion utilities now live in the `arkrpc/treeconv`
+subpackage to keep the generated client package lighter.
 Proto source: `arkrpc/ark.proto`, `arkrpc/indexer.proto`.
 
 ## Key Types
 
 - `TreePath` / `TreePathNode` / `TxOut` — Structured proto messages for the
   VTXO commitment tree path. Replaces opaque TLV bytes on the wire.
-- `TreePathFromTree` / `TreePathToTree` — Lossless conversion between
-  `tree.Tree` and `arkrpc.TreePath`. Uses deterministic pre-order flattening
-  with sorted child indices.
 - `IncomingOOREvent` — Lightweight notification (wake-up hint). Carries only
   session_id, pk_script, event_id. Triggers the three-phase receive flow.
 - `OORRecipientEvent` — Phase 1 query response from
@@ -23,14 +21,12 @@ Proto source: `arkrpc/ark.proto`, `arkrpc/indexer.proto`.
 
 ## Relationships
 
-- **Depends on**: `lib/tree` (for conversion utilities in `tree_path_convert.go`).
+- **Depends on**: generated gRPC types only.
 - **Depended on by**: `indexer`, `darepod`, `serverconn`, `oor` (uses generated
-  clients and conversion helpers).
+  clients). Tree conversion helpers are consumed via `arkrpc/treeconv`.
 
 ## Invariants
 
 - **Never edit generated code** — regenerate via `make rpc`.
-- Conversion round-trip: `TreePathFromTree(t)` → `TreePathToTree(pb)` must
-  reproduce the original tree (excluding derived `FinalKey` fields).
 - Child iteration during flattening is sorted by output index for
   deterministic serialization.
